requirements.txt
asgiref==3.4.1
autocommand==2.2.1
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.2.4
Django==4.0.1
django-allauth==0.47.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
filelock==3.4.2
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
flast==1.3
fonttools==4.31.2
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.1.1
packaging==21.3
path==16.4.0
Pillow==9.0.0
pip-run==8.8.0
pipenv==2022.1.8
platformdirs==2.4.1
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
Werkzeug==2.1.2
